Fix review discussions

+++ cloud/disk_manager/internal/pkg/dataplane/filesystem/traversal/storage/storage.go
@@ -44,6 +44,6 @@ type Storage interface {
 	ClearDirectoryListingQueue(
 		ctx context.Context,
 		snapshotID string,
-		deletionLimit int,
+		deletionLimit uint64,
 	) error
 }
